Refactor service interfaces
