+++ xwe/services/combat_service.py
@@ -3,46 +3,11 @@
 负责战斗系统的逻辑处理
 """
 
-from abc import ABC, abstractmethod
 from typing import Any, Dict, List, Optional
-from dataclasses import dataclass
 
 from xwe.services import ServiceBase, ServiceContainer
 from xwe.services.game_service import CommandResult
-
-
-class ICombatService(ABC):
-    """战斗服务接口"""
-
-    @abstractmethod
-    def start_combat(self, enemy_data: Dict[str, Any]) -> bool:
-        """开始战斗"""
-        pass
-
-    @abstractmethod
-    def execute_attack(self, target: Optional[str] = None) -> CommandResult:
-        """执行攻击"""
-        pass
-
-    @abstractmethod
-    def execute_defend(self) -> CommandResult:
-        """执行防御"""
-        pass
-
-    @abstractmethod
-    def attempt_flee(self) -> CommandResult:
-        """尝试逃跑"""
-        pass
-
-    @abstractmethod
-    def is_in_combat(self) -> bool:
-        """是否在战斗中"""
-        pass
-
-    @abstractmethod
-    def get_combat_state(self) -> Dict[str, Any]:
-        """获取战斗状态"""
-        pass
+from xwe.services.interfaces.combat_service import ICombatService
 
 
 class CombatService(ServiceBase[ICombatService], ICombatService):
